backend/requirements.txt
# Django核心
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0

# 环境变量
python-dotenv==1.0.1

# 数据库
pymysql==1.1.2

# WebSocket
channels==4.3.2
channels-redis==4.2.1

# Celery异步任务
celery==5.6.2
redis==7.1.0

# 缓存
django-redis==6.0.0

# 数据处理
pandas==2.2.3
openpyxl==3.1.5

# 安全
cryptography==46.0.3

# 图片处理
Pillow==12.1.0
